scheduler/requirements.txt
backoff
pika
psycopg2-binary
pydantic
backoff
requests
